requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0
matplotlib>=3.5.0
MetaTrader5>=5.0.34
PyYAML>=6.0
pyttsx3>=2.90
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
pytest>=6.2.5
dask>=2021.10.0
pyspark>=3.1.2
boto3>=1.18.0
google-cloud-aiplatform>=1.7.0
transformers>=4.11.0
requests>=2.25.1
ccxt>=4.0.0
websocket-client>=1.6.0
python-binance>=1.0.19
TA-Lib>=0.4.25
fastapi>=0.104.0
uvicorn>=0.24.0
redis>=5.0.0
aiofiles>=23.2.0
asyncio-mqtt>=0.16.0
